requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
python-dotenv==1.0.0
google-generativeai==0.3.1
SQLAlchemy==2.0.23
werkzeug==3.0.1
alembic==1.13.0
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.4
youtube-transcript-api==0.6.1
markdown==3.5.1
beautifulsoup4==4.12.2
bleach==6.1.0
gunicorn==21.2.0
pytest==7.4.3
